refactor(common): move task-hub binding into a telemetry submodule
bind_task_hub, the AdoptMainClient poll combinator and the wasm passthrough
move verbatim out of lib.rs into telemetry.rs. bind_task_hub stays re-exported
at the crate root: it is what the span macros emit and what call sites use.
